split ShardManager from RoutingFilter
